# docs/specs/纵行科技_智能运维Agent_比赛Demo_API_v1.openapi.yaml
openapi: 3.1.0
info:
  title: ZiFiSense Intelligent Maintenance Agent API
  version: 1.0.0
  description: >-
    Competition evaluation API for the ZiFiSense intelligent maintenance Agent.
    The API exposes an isolated evaluation session, natural-language Agent invocation,
    industrial event ingestion, task inspection, and approval-gated work-order actions.
servers:
  - url: https://agent-demo.example.com
    description: Competition deployment placeholder
...
paths:
...
  /api/v1/capabilities:
    get:
      tags: [Evaluation]
      operationId: getCapabilities
      summary: Discover testable Agent capabilities and scenarios
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Capability catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/evaluation/sessions:
    post:
      tags: [Evaluation]
      operationId: createEvaluationSession
      summary: Create an isolated evaluation session and load the initial alarm
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvaluationSessionRequest"
      responses:
        "201":
          description: Evaluation session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEvaluationSessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/tasks/{task_id}:
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Get task state, evidence, conflicts, actions, and validation result
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSnapshotResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tasks/{task_id}/approvals:
    post:
      tags: [Tasks]
      operationId: decideApproval
      summary: Approve or reject a pending work-order action
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Approval decision applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/admin/reset:
    post:
      tags: [Admin]
      operationId: resetDemoData
      summary: Reset one evaluation session or all competition fixtures
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetRequest"
      responses:
        "200":
          description: Reset completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque API Key
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-provided request identifier; the server generates one when omitted.
      schema:
        type: string
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key for safely retrying a write operation.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Valid identity without the required scope or object access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: State, evidence version, or idempotency conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [api_version, timestamp, is_degraded]
      properties:
        api_version:
          type: string
          const: v1
        timestamp:
          type: string
          format: date-time
        is_degraded:
          type: boolean
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [request_id, trace_id, error]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - INVALID_ACCESS_TOKEN
                - INSUFFICIENT_SCOPE
                - RESOURCE_NOT_FOUND
                - INVALID_STATE_TRANSITION
                - EVIDENCE_VERSION_CONFLICT
                - IDEMPOTENCY_CONFLICT
                - APPROVAL_CHALLENGE_INVALID
                - RATE_LIMITED
                - MODEL_UNAVAILABLE
                - SERVICE_UNAVAILABLE
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
    CapabilitiesResponse:
      type: object
      additionalProperties: false
      required: [request_id, trace_id, data, meta]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        data:
          $ref: "#/components/schemas/CapabilitiesData"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    CapabilitiesData:
      type: object
      additionalProperties: false
      required: [agent, scenarios, capabilities, supported_event_types, safety_boundaries]
      properties:
        agent:
          type: object
          additionalProperties: false
          required: [name, version, locale]
          properties:
            name:
              type: string
              example: zifisense-intelligent-maintenance-agent
            version:
              type: string
              example: 1.0.0
            locale:
              type: string
              example: zh-CN
        scenarios:
          type: array
          items:
            $ref: "#/components/schemas/ScenarioSummary"
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/CapabilityItem"
        supported_event_types:
          type: array
          items:
            $ref: "#/components/schemas/EventType"
        safety_boundaries:
          type: array
          items:
            type: string
          example:
            - LLM does not diagnose raw high-frequency time-series data.
            - Production-control actions are not exposed.
            - Formal work-order submission requires explicit approval.
    ScenarioSummary:
      type: object
      additionalProperties: false
      required: [scenario_id, name, description, suggested_questions]
      properties:
        scenario_id:
          type: string
          example: reducer_gear_alarm_v1
        name:
          type: string
          example: Main-drive reducer alarm investigation
        description:
          type: string
        suggested_questions:
          type: array
          items:
            type: string
    CapabilityItem:
      type: object
      additionalProperties: false
      required: [code, name, verification_hint]
      properties:
        code:
          type: string
          enum:
            - asset_catalog_query
            - current_fault_query
            - fault_detail_query
            - fault_history_query
            - monitoring_summary
            - operating_context_query
            - maintenance_history_query
            - peer_comparison
            - knowledge_retrieval
            - context_orchestration
            - evidence_conflict_detection
            - human_input_gate
            - portable_measurement_ingest
            - work_order_draft
            - approval_gated_write
            - maintenance_result_validation
            - audit_traceability
        name:
          type: string
        verification_hint:
          type: string
    CreateEvaluationSessionRequest:
      type: object
      additionalProperties: false
      required: [scenario_id]
      properties:
        scenario_id:
          type: string
          const: reducer_gear_alarm_v1
        locale:
          type: string
          default: zh-CN
    CreateEvaluationSessionResponse:
      type: object
      additionalProperties: false
      required: [request_id, trace_id, data, meta]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        data:
          type: object
          additionalProperties: false
          required:
            - evaluation_session_id
            - conversation_id
            - task_id
            - scenario_id
            - task_state
            - scenario_summary
            - suggested_questions
          properties:
            evaluation_session_id:
              type: string
            conversation_id:
              type: string
            task_id:
              type: string
            scenario_id:
              type: string
            task_state:
              $ref: "#/components/schemas/TaskState"
            scenario_summary:
              type: string
            suggested_questions:
              type: array
              items:
                type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    EvidenceItem:
      type: object
      additionalProperties: false
      required:
        - evidence_id
        - evidence_type
        - summary
        - source_system
        - observed_at
        - quality_status
        - usage_level
        - is_simulated
      properties:
        evidence_id:
          type: string
        evidence_type:
          type: string
          enum: [ALARM, KNOWLEDGE, OPERATING_CONTEXT, MAINTENANCE_HISTORY, HUMAN_CLAIM, PORTABLE_MEASUREMENT, WORK_ORDER_RESULT]
        summary:
          type: string
        source_system:
          type: string
        observed_at:
          type: string
          format: date-time
        quality_status:
          type: string
          enum: [VALID, INCOMPLETE, STALE, CONFLICTING, UNVERIFIED, REJECTED]
        usage_level:
          type: string
          enum: [RECORD_ONLY, QUERY_GUIDANCE, DECISION_REFERENCE, VERIFIED_LABEL]
        is_simulated:
          type: boolean
        conflicts_with:
          type: array
          items:
            type: string
...
    ToolExecution:
      type: object
      additionalProperties: false
      required: [tool_name, status, source_system, elapsed_ms]
      properties:
        tool_name:
          type: string
        status:
          type: string
          enum: [SUCCEEDED, FAILED, TIMED_OUT, SKIPPED]
        source_system:
          type: string
        elapsed_ms:
          type: integer
          minimum: 0
        is_simulated:
          type: boolean
...
    PendingApproval:
      type: object
      additionalProperties: false
      required: [approval_id, approval_challenge, approval_type, evidence_version, expires_at]
      properties:
        approval_id:
          type: string
        approval_challenge:
          type: string
        approval_type:
          type: string
          enum: [SUBMIT_WORK_ORDER]
        evidence_version:
          type: integer
          minimum: 1
        expires_at:
          type: string
          format: date-time
        impact_preview:
          type: object
          additionalProperties: true
...
    EventType:
      type: string
      enum: [ALARM_RAISED, FIELD_MEASUREMENT_COMPLETED, WORK_ORDER_COMPLETED]
    AlarmRaisedPayload:
      type: object
      additionalProperties: false
      required: [alarm_id, asset_id, measurement_point_id, severity, diagnosis_text, confidence, algorithm_version]
      properties:
        alarm_id:
          type: string
        asset_id:
          type: string
        measurement_point_id:
          type: string
        severity:
          type: string
          enum: [INFO, WARNING, CRITICAL]
        diagnosis_text:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        algorithm_version:
          type: string
        evidence_features:
          type: array
          items:
            type: object
            additionalProperties: true
...
    TaskSnapshotResponse:
      type: object
      additionalProperties: false
      required: [request_id, trace_id, data, meta]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        data:
          $ref: "#/components/schemas/TaskSnapshot"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    TaskSnapshot:
      type: object
      additionalProperties: false
      required:
        - task_id
        - evaluation_session_id
        - task_state
        - evidence_version
        - evidence
        - conflicts
        - tool_executions
        - timeline
      properties:
        task_id:
          type: string
        evaluation_session_id:
          type: string
        task_state:
          $ref: "#/components/schemas/TaskState"
        evidence_version:
          type: integer
          minimum: 1
        alarm:
          $ref: "#/components/schemas/AlarmRaisedPayload"
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceConflict"
        tool_executions:
          type: array
          items:
            $ref: "#/components/schemas/ToolExecution"
        pending_approval:
          oneOf:
            - $ref: "#/components/schemas/PendingApproval"
            - type: "null"
        work_order:
          oneOf:
            - $ref: "#/components/schemas/WorkOrderSummary"
            - type: "null"
        maintenance_validation:
          oneOf:
            - $ref: "#/components/schemas/MaintenanceValidation"
            - type: "null"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
    EvidenceConflict:
      type: object
      additionalProperties: false
      required: [conflict_id, summary, evidence_ids, status]
      properties:
        conflict_id:
          type: string
        summary:
          type: string
        evidence_ids:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [OPEN, RESOLVED, ACCEPTED_UNCERTAINTY]
    WorkOrderSummary:
      type: object
      additionalProperties: false
      required: [work_order_id, status, title]
      properties:
        work_order_id:
          type: string
        status:
          type: string
          enum: [DRAFT, APPROVAL_PENDING, SUBMITTED, IN_PROGRESS, COMPLETED]
        title:
          type: string
        recommended_window:
          type: string
    MaintenanceValidation:
      type: object
      additionalProperties: false
      required: [status, actual_fault, sample_status]
      properties:
        status:
          type: string
          enum: [PENDING, ENGINEER_CONFIRMED, VERIFIED, CONFLICTING]
        actual_fault:
          type: string
        sample_status:
          type: string
          enum: [NOT_ELIGIBLE, PENDING_REVIEW, APPROVED]
        summary:
          type: string
    TimelineEvent:
      type: object
      additionalProperties: false
      required: [event_id, event_type, occurred_at, summary]
      properties:
        event_id:
          type: string
        event_type:
          type: string
        occurred_at:
          type: string
          format: date-time
        summary:
          type: string
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required: [approval_id, approval_challenge, decision, evidence_version]
      properties:
        approval_id:
          type: string
        approval_challenge:
          type: string
        decision:
          type: string
          enum: [APPROVE, REJECT]
        evidence_version:
          type: integer
          minimum: 1
        comment:
          type: string
          maxLength: 2000
    ApprovalDecisionResponse:
      type: object
      additionalProperties: false
      required: [request_id, trace_id, data, meta]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        data:
          type: object
          additionalProperties: false
          required: [approval_id, decision, task_id, task_state]
          properties:
            approval_id:
              type: string
            decision:
              type: string
              enum: [APPROVE, REJECT]
            task_id:
              type: string
            task_state:
              $ref: "#/components/schemas/TaskState"
            work_order_id:
              type: string
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ResetRequest:
      type: object
      additionalProperties: false
      required: [scope]
      properties:
        scope:
          type: string
          enum: [SESSION, ALL_FIXTURES]
        evaluation_session_id:
          type: string
    ResetResponse:
      type: object
      additionalProperties: false
      required: [request_id, trace_id, data, meta]
      properties:
        request_id:
          type: string
        trace_id:
          type: string
        data:
          type: object
          additionalProperties: false
          required: [scope, status, reset_count]
          properties:
            scope:
              type: string
            status:
              type: string
              const: COMPLETED
            reset_count:
              type: integer
              minimum: 0
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    TaskState:
      type: string
      enum:
        - ALARM_RECEIVED
        - CONTEXT_COLLECTING
        - EVIDENCE_REVIEW
        - FIELD_EVIDENCE_PENDING
        - HUMAN_DECISION
        - WORK_ORDER_DRAFTED
        - APPROVAL_PENDING
        - MAINTENANCE_PROCESSING
        - RESULT_VALIDATION
        - CLOSED